improve documentation and file structure
