+++ wrappers/python/wpwithin_python/wpwithin_python/launcher.py
@@ -1,9 +1,14 @@
+"""
+Launcher for RPC Agent.
+File name should be rpc-agent-[system]-[architecture]. ie rpc-agent-mac-64
+"""
+
 from subprocess import Popen, PIPE
 import sys
 import platform
 
 def os_arch():
-    """Give the architecture as 32 or 64"""
+    """Give the architecture as '32' or '64'"""
     if platform.machine().lower()[:3] == 'arm':
         out = 'arm'
     elif sys.maxsize > 2**32:
@@ -13,7 +18,7 @@ def os_arch():
     return out
 
 def os_platform():
-    """Give the platform as win, mac or linux"""
+    """Give the platform as 'win', 'mac' or 'linux'"""
     os = platform.system().lower()
     if os == 'windows':
         out = 'win'
@@ -26,8 +31,9 @@ def os_platform():
 def run_rpc_agent(port, rpc_dir, callback_port=None):
     """Run RPC Agent
     Args:
-        execPath (string): path to directory with rpc agent launchers
         port (integer): port to run RPC agent on
+        rpc_dir (string): path to directory with rpc agent launchers
+        (optional) callback_port (integer): port to listen for callback events
     """
     os = os_platform()
     agent = 'rpc-agent-' + os + '-' + os_arch()
