+++ wrappers/python/wpwithin_python/wpwithin_python/wpwithin_callbacks.py
@@ -1,3 +1,8 @@
+"""Abstract class for event listeners to inherit and implement.
+For examples, see
+https://github.com/WPTechInnovation/worldpay-within-sdk/tree/feature/python-wrapper/wrappers/python/wpwithin_python/examples
+"""
+
 class AbstractEventListener(object):
     """Abstract class with methods required by WPWithin Callback Service.
     For examples, see
